requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\athey\Documents\Linktame\linktame_api_main.py: 18
Flask == 2.0.1

# C:\Users\athey\Documents\Linktame\linktame_api_main.py: 22
Flask_Cors == 3.0.10

# C:\Users\athey\Documents\Linktame\linktame_api_main.py: 19
Flask_SQLAlchemy == 2.5.1

# C:\Users\athey\Documents\Linktame\linktame_api_main.py: 14
PyJWT == 2.1.0

# C:\Users\athey\Documents\Linktame\linktame_api_main.py: 11
Werkzeug == 2.0.1

# C:\Users\athey\Documents\Linktame\linktame_api_main.py: 5
gunicorn == 20.1.0

# C:\Users\athey\Documents\Linktame\linktame_api_main.py: 20
psycopg2_binary == 2.9.1

# C:\Users\athey\Documents\Linktame\linktame_api_main.py: 7
requests == 2.25.1
